// swiyu-trust-management-ui/src/app/api/generated/model/page-metadata.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface PageMetadata {
  size?: number;
  number?: number;
  totalElements?: number;
  totalPages?: number;
}