// swiyu-trust-management-ui/src/app/api/generated/model/trust-onboarding-submission-document-list-item-dto.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface TrustOnboardingSubmissionDocumentListItemDto {
  id?: string;
  name?: string;
  mediaType?: string;
  type?: string;
  owningBusinessPartner?: string;
  createdAt?: string;
  updatedAt?: string;
  submittedAt?: string;
  trustOnboardingSubmissionId?: string;
}